+++ packages/skeleton-svelte/src/routes/components/input-chips/+page.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
-	import { TagsInput } from '$lib/index.js';
+	import { InputChips } from '$lib/index.js';
 
 	let value = $state(['red', 'green', 'blue']);
 </script>
 
 <div class="space-y-10">
 	<header>
-		<h1 class="h1">Tags Input</h1>
+		<h1 class="h1">Input Chips</h1>
 	</header>
 	<section class="space-y-4">
 		<pre class="pre">{JSON.stringify(value, null, 2)}</pre>
-		<TagsInput bind:value></TagsInput>
+		<InputChips bind:value></InputChips>
 	</section>
 </div>
